cleaning personel menu

+++ src/Accountant.java
@@ -1,6 +1,6 @@
 public class Accountant extends Employee {
 
-    public Accountant(String name, int phoneNum, double salary) {
+    public Accountant(String name, int phoneNum, int salary) {
         super(name, phoneNum, salary);
     }
