+++ src/Director.java
@@ -2,7 +2,7 @@ public class Director extends Employee {
 
     // The constructor is not needed if we are not adding additional info specific
     // for the director
-    public Director(String name, int phoneNum, double salary) {
+    public Director(String name, int phoneNum, int salary) {
         super(name, phoneNum, salary);
     }
